+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -97,6 +97,12 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`cppcoreguidelines-avoid-capture-this-with-capture-default
+  <clang-tidy/checks/cppcoreguidelines/avoid-capture-this-with-capture-default>` check.
+
+  Warns when lambda specify a capture default and capture ``this``. The check also
+  offers FixIts.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^